PAYOSWXP-130: get rid of duplicate code
